requirements.txt
nltk
torch
torchvision
pillow
streamlit
pytest